refactor: clippy cleanup and builder
