+++ .github/workflows/ci.yaml
@@ -20,9 +20,9 @@ jobs:
           components: rustfmt, clippy
       - name: cargo check
         run: cargo check --all --tests
-      - name: cargo clippy
-        run: cargo clippy -- -D warnings
       - name: cargo fmt
         run: cargo fmt -- --check
+      - name: cargo clippy
+        run: cargo clippy -- -D warnings
       - name: cargo test
         run: cargo test --all
