+++ scroll_core/src/archive/archive_loader.rs
@@ -1,6 +1,5 @@
 //    archive_loader.rs
 //======================================
-#![allow(unused_imports)]
 
 use crate::parser;
 use crate::scroll::Scroll;
