+++ scroll_core/src/artifact.rs
@@ -1,6 +1,5 @@
 // artifact.rs – Abstract Writable Artifact Interface
 //======================================================
-#![allow(unused_imports)]
 
 pub trait WritableArtifact {
     fn to_string_representation(&self) -> String;
