+++ scroll_core/src/core/cost_manager.rs
@@ -1,12 +1,10 @@
 // cost_manager.rs – The Core Weave
 //====================================
-#![allow(unused_imports)]
 
 use crate::errors::MetricError;
-use crate::invocation::invocation_core::Invocation;
+use crate::invocation::types::Invocation;
 use crate::metrics::clamp_finite;
 use crate::scroll::Scroll;
-use chrono::Utc;
 
 #[cfg(test)]
 use std::cell::RefCell;
