+++ scroll_core/src/invocation/constructs/file_reader_construct.rs
@@ -1,4 +1,4 @@
-use crate::invocation::invocation_core::{Invocation, InvocationResult};
+use crate::invocation::types::{Invocation, InvocationResult};
 use crate::invocation::named_construct::NamedConstruct;
 use crate::orchestra::{AgentMessage, Bus, OrchestratedConstruct};
 use serde_json::json;
@@ -21,7 +21,7 @@ impl NamedConstruct for FileReader {
         _invocation: &Invocation,
         _scroll: Option<crate::Scroll>,
     ) -> Result<InvocationResult, String> {
-        Ok(InvocationResult::Success("ready".into()))
+        Ok(InvocationResult::Success(Box::from("ready")))
     }
 }
