+++ scroll_core/src/invocation/mod.rs
@@ -4,7 +4,7 @@
 
 pub mod aelren;
 pub mod constructs;
-pub mod invocation_core;
+pub mod types;
 pub mod invocation_manager;
 pub mod ledger;
 pub mod named_construct;
