+++ scroll_core/src/runner/mod.rs
@@ -1,2 +1,2 @@
 pub mod invocation_context;
-pub mod runner_core;
+pub mod core;
