+++ scroll_core/src/scroll_writer.rs
@@ -1,6 +1,5 @@
 // scroll_writer.rs – Hand of the Archive
 //===========================================
-#![allow(unused_imports)]
 
 use chrono::Utc;
 use serde_yaml;
