+++ scroll_core/src/trigger_loom/engine.rs
@@ -11,7 +11,7 @@ use log::info;
 use uuid::Uuid;
 
 use crate::core::cost_manager::{CostDecision, CostManager, InvocationCost};
-use crate::invocation::invocation_core::{Invocation, InvocationMode, InvocationTier};
+use crate::invocation::types::{Invocation, InvocationMode, InvocationTier};
 use crate::invocation::named_construct::NamedConstruct;
 use crate::trigger_loom::config::TriggerLoopConfig;
