+++ scroll_core/src/trigger_loom/loom.rs
@@ -2,7 +2,7 @@
 // src/trigger_loom/loom.rs
 // ===============================
 
-use crate::invocation::invocation_core::Invocation;
+use crate::invocation::types::Invocation;
 use crate::invocation::named_construct::NamedConstruct;
 use crate::trigger_loom::emotional_state::EmotionalState;
 use crate::trigger_loom::glyph_matcher::{match_glyph, GlyphMatchResult};
